refactor(scripts): extract shared backfill apply-phase helper
SonarCloud flagged 38% duplication on PR #919's new code — the pool/prisma
lifecycle + source lookup + fingerprint dedup + createMany block was lifted
verbatim from backfill-bfmh3-history.ts. Every future backfill script would
accumulate the same block, so extract it once now rather than retrofitting
the 13 existing scripts in the same PR.
